fix: Added support for no auth based on BE
